Move GraphicDetails into rendering package
